rename layers in folders
