+++ services/web/server/src/simcore_service_webserver/workspaces/_trash_api.py
@@ -10,7 +10,7 @@ from models_library.workspaces import WorkspaceID, WorkspaceUpdateDB
 from simcore_postgres_database.utils_repos import transaction_context
 
 from ..db.plugin import get_asyncpg_engine
-from ..folders._trash_api import trash_folder, untrash_folder
+from ..folders._trash_service import trash_folder, untrash_folder
 from ..projects._trash_service import trash_project, untrash_project
 from ._workspaces_api import check_user_workspace_access
 from ._workspaces_db import update_workspace
